Added basic retrival of API data
Made a basic sample of how data should/could be retrieved from the API
using Agfx (recent posts for now).
This includes mapping from json to POC and binding in XAML so we can
actually see stuff in the app.
Should give you idea of how data is retrieved in the app.